Add IdMapping to fix calling contract on MyCrypto.com (#33)
